Update code_basic.py
This version of the code is more computationally friendly, as it does not iterate over each row and column individually, but it chooses an imputation strategy at the beginning (based on Q learning) and then applies it to the entire dataset for the given episode. Because the agent will only have the number of episodes to learn the optimal strategy, it will converge quicker but be less precise than the advanced code version.